Fix incorrect use of FLT_MIN
Fix incorrect use of FLT_MIN (should be - FLT_MAX), affecting HDR
dynamic range calculation and compression of HDR half-float RGB(A)
format textures.

Some other minor changes.

+++ rgtc.c
@@ -110,7 +110,7 @@ int offset, int flags) {
 			case 7 : result = 127; break;
 			}
 		// Map from [-127, 127] to [-32768, 32767].
-		*((uint16_t *)&image_buffer[i] + offset) = (uint16_t)(int16_t)((result + 127) * 65535 / 254 - 32768);
+		*((int16_t *)&image_buffer[i] + offset) = (int16_t)((result + 127) * 65535 / 254 - 32768);
 		bits >>= 3;
 	}
 	return 1;
